+++ submit_cgap/scripts/submit_metadata_bundle.py
@@ -44,8 +44,7 @@ def main(simulated_args_for_testing=None):
                              server=args.server, env=args.env,
                              validate_only=args.validate_only, upload_folder=args.upload_folder,
                              no_query=args.no_query, subfolders=args.subfolders, app=args.app,
-                             submission_protocol=args.submission_protocol,
-                             )
+                             submission_protocol=args.submission_protocol)
 
 
 if __name__ == '__main__':