+++ submit_cgap/tests/test_submission.py
@@ -298,8 +298,8 @@ def make_user_record(title='J Doe',
 def test_get_user_record():
 
     def make_mocked_get(auth_failure_code=400):
-        def mocked_get(url, *, auth):
-            ignored(url)
+        def mocked_get(url, *, auth, **kwargs):
+            ignored(url, kwargs)
             if auth != SOME_AUTH:
                 return FakeResponse(status_code=auth_failure_code, json={'Title': 'Not logged in.'})
             return FakeResponse(status_code=200, json={'title': 'J Doe', 'contact_email': 'jdoe@cgap.hms.harvard.edu'})
@@ -420,7 +420,7 @@ def test_show_upload_info():
 
     json_result = None  # Actual value comes later
 
-    def mocked_get(url, *, auth):
+    def mocked_get(url, *, auth, **kwargs):
         assert url.startswith(SOME_UUID_UPLOAD_URL)
         assert auth == SOME_AUTH
         return FakeResponse(200, json=json_result)
@@ -431,7 +431,7 @@ def test_show_upload_info():
             json_result = {}
             with shown_output() as shown:
                 show_upload_info(SOME_UUID, server=SOME_SERVER, env=None, keydict=SOME_KEYDICT)
-                assert shown.lines == ['No uploads.']
+                assert shown.lines == ['Uploads: None']
 
             json_result = SOME_UPLOAD_INFO_RESULT
             with shown_output() as shown:
@@ -617,7 +617,7 @@ def test_do_any_uploads():
                     auth=SOME_KEYDICT,
                     folder=SOME_BUNDLE_FILENAME_FOLDER,  # the folder part of given SOME_BUNDLE_FILENAME
                     no_query=False,
-                    subfolders=False,
+                    subfolders=False
                 )
                 assert shown.lines == []
 
@@ -633,7 +633,7 @@ def test_do_any_uploads():
                     auth=SOME_KEYDICT,
                     folder=SOME_OTHER_BUNDLE_FOLDER,  # passed straight through
                     no_query=False,
-                    subfolders=False,
+                    subfolders=False
                 )
                 assert shown.lines == []
 
@@ -649,7 +649,7 @@ def test_do_any_uploads():
                     auth=SOME_KEYDICT,
                     folder=None,  # No folder
                     no_query=False,
-                    subfolders=False,
+                    subfolders=False
                 )
                 assert shown.lines == []
 
@@ -666,7 +666,7 @@ def test_do_any_uploads():
                     auth=SOME_KEYDICT,
                     folder=SOME_BUNDLE_FILENAME_FOLDER,  # the folder part of given SOME_BUNDLE_FILENAME
                     no_query=False,
-                    subfolders=True,
+                    subfolders=True
                 )
                 assert shown.lines == []
 
@@ -679,14 +679,14 @@ def test_do_any_uploads():
                 res=SOME_UPLOAD_INFO_RESULT,
                 keydict=SOME_KEYDICT,
                 ingestion_filename=SOME_BUNDLE_FILENAME,  # from which a folder can be inferred
-                no_query=True,
+                no_query=True
             )
             mock_uploads.assert_called_with(
                 SOME_UPLOAD_INFO,
                 auth=SOME_KEYDICT,
                 folder=SOME_BUNDLE_FILENAME_FOLDER,  # the folder part of given SOME_BUNDLE_FILENAME
                 no_query=True,
-                subfolders=False,
+                subfolders=False
             )
             assert shown.lines == []
 
@@ -707,7 +707,7 @@ def test_resume_uploads():
                             ingestion_filename=SOME_BUNDLE_FILENAME,
                             upload_folder=None,
                             no_query=False,
-                            subfolders=False,
+                            subfolders=False
                         )
 
     with mock.patch.object(utils_module, "script_catch_errors", script_dont_catch_errors):
@@ -783,8 +783,9 @@ def test_execute_prearranged_upload(os_simulation_mode: str):
                             **subprocess_options
                         )
                         assert shown.lines == [
-                            "Going to upload some-filename to some-url.",
-                            "Uploaded in 1.00 seconds"  # 1 tick (at rate of 1 second per tick in our controlled time)
+                            "Uploading local file some-filename directly (via AWS CLI) to: some-url",
+                            # 1 tick (at rate of 1 second per tick in our controlled time)
+                            "Upload duration: 1.00 seconds"
                         ]
 
         with mock.patch.object(os, "environ", SOME_ENVIRON.copy()):
@@ -801,8 +802,9 @@ def test_execute_prearranged_upload(os_simulation_mode: str):
                             **subprocess_options
                         )
                         assert shown.lines == [
-                            "Going to upload some-filename to some-url.",
-                            "Uploaded in 1.00 seconds"  # 1 tick (at rate of 1 second per tick in our controlled time)
+                            "Uploading local file some-filename directly (via AWS CLI) to: some-url",
+                            # 1 tick (at rate of 1 second per tick in our controlled time)
+                            "Upload duration: 1.00 seconds"
                         ]
 
         with mock.patch.object(os, "environ", SOME_ENVIRON.copy()):
@@ -817,7 +819,7 @@ def test_execute_prearranged_upload(os_simulation_mode: str):
                             **subprocess_options
                         )
                         assert shown.lines == [
-                            "Going to upload some-filename to some-url.",
+                            "Uploading local file some-filename directly (via AWS CLI) to: some-url",
                         ]
 
 
@@ -1043,7 +1045,7 @@ def test_do_uploads(tmp_path):
         mock_upload.assert_called_with(
             filename=file_path,
             uuid=uuid,
-            auth=SOME_AUTH,
+            auth=SOME_AUTH
         )
 
     with mock.patch.object(submission_module, "upload_file_to_uuid") as mock_upload:
@@ -1057,7 +1059,7 @@ def test_do_uploads(tmp_path):
         mock_upload.assert_called_with(
             filename=(folder.as_posix() + "/" + filename),
             uuid=uuid,
-            auth=SOME_AUTH,
+            auth=SOME_AUTH
         )
 
     with mock.patch.object(submission_module, "upload_file_to_uuid") as mock_upload:
@@ -1072,7 +1074,7 @@ def test_do_uploads(tmp_path):
         mock_upload.assert_called_with(
             filename=file_path,
             uuid=uuid,
-            auth=SOME_AUTH,
+            auth=SOME_AUTH
         )
 
     with mock.patch.object(submission_module, "upload_file_to_uuid") as mock_upload:
@@ -1135,7 +1137,7 @@ def test_do_uploads(tmp_path):
                         mocked_instance,
                         folder,
                         SOME_AUTH,
-                        recursive=False,
+                        recursive=False
                     )
 
 
@@ -1146,6 +1148,7 @@ def test_upload_item_data():
             with mock.patch.object(submission_module, "yes_or_no", return_value=True):
                 with mock.patch.object(submission_module, "upload_file_to_uuid") as mock_upload:
 
+                    pass
                     upload_item_data(item_filename=SOME_FILENAME, uuid=SOME_UUID, server=SOME_SERVER, env=SOME_ENV)
 
                     mock_resolve.assert_called_with(env=SOME_ENV, server=SOME_SERVER)
@@ -1182,8 +1185,7 @@ def test_upload_item_data():
 
                 mock_resolve.assert_called_with(env=SOME_ENV, server=SOME_SERVER)
                 mock_get.assert_called_with(SOME_SERVER)
-                mock_upload.assert_called_with(filename=SOME_FILENAME, uuid=SOME_UUID,
-                                               auth=SOME_KEYDICT)
+                mock_upload.assert_called_with(filename=SOME_FILENAME, uuid=SOME_UUID, auth=SOME_KEYDICT)
 
 
 def get_today_datetime_for_time(time_to_use):
@@ -1214,37 +1216,69 @@ class Scenario:
     def make_uploaded_lines(self):
         uploaded_time = self.get_time_after_wait()
         return [
-            f"The server {SOME_SERVER} recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.",
+            f"The server {SOME_SERVER} recognizes you as: J Doe <jdoe@cgap.hms.harvard.edu>",
             (
-                f"{uploaded_time} Bundle uploaded, assigned uuid {SOME_UUID} for tracking."
-                " Awaiting processing..."
+                (f"{uploaded_time} Bundle uploaded."
+                 f" Checking ingestion process using IngestionSubmission uuid: {SOME_UUID} ...")
             ),
         ]
 
-    def make_wait_lines(self, wait_attempts):
+    def make_wait_lines(self, wait_attempts, outcome: str = None):
         result = []
         uploaded_time = self.get_time_after_wait()
-        for idx in range(wait_attempts):
-            time_delta_from_start = (PROGRESS_CHECK_INTERVAL + self.wait_time_delta) * (idx + 1)
+        time_delta_from_start = 0
+        nchecks = 0
+        ERASE_LINE = "\033[K"
+        for idx in range(wait_attempts + 1):
+            time_delta_from_start += 1
             adjusted_scenario = Scenario(start_time=uploaded_time, wait_time_delta=time_delta_from_start)
             wait_time = adjusted_scenario.get_time_after_wait()
-            wait_line = f"{wait_time} Progress is not done yet. Continuing to wait..."
+            wait_line = (f"{ERASE_LINE}{wait_time} Checking processing"
+                         f" | Status: Not Done Yet | Checked: {nchecks} time{'s' if nchecks != 1 else ''} ...\r")
             result.append(wait_line)
+            if nchecks >= wait_attempts:
+                time_delta_from_start += 1
+                adjusted_scenario = Scenario(start_time=uploaded_time, wait_time_delta=time_delta_from_start)
+                wait_time = adjusted_scenario.get_time_after_wait()
+                if outcome == "timeout":
+                    wait_line = (f"{ERASE_LINE}{wait_time} Giving up waiting for processing completion"
+                                 f" | Status: Not Done Yet | Checked: {nchecks + 1} times\n\r")
+                else:
+                    wait_line = (f"{ERASE_LINE}{wait_time} Processing complete"
+                                 f" | Status: {outcome.title() if outcome else 'Unknown'}"
+                                 f" | Checked: {nchecks + 1} times\n\r")
+                result.append(wait_line)
+                break
+            nchecks += 1
+            for i in range(PROGRESS_CHECK_INTERVAL):
+                time_delta_from_start += 2  # Extra 1 for the 1-second sleep loop in utils.check_repeatedly
+                adjusted_scenario = Scenario(start_time=uploaded_time, wait_time_delta=time_delta_from_start)
+                wait_time = adjusted_scenario.get_time_after_wait()
+                wait_line = (
+                    f"{ERASE_LINE}{wait_time} Waiting for processing completion"
+                    f" | Status: Not Done Yet | Checked: {idx + 1} time{'s' if idx + 1 != 1 else ''}"
+                    f" | Next check: {PROGRESS_CHECK_INTERVAL - i}"
+                    f" second{'s' if PROGRESS_CHECK_INTERVAL - i != 1 else ''} ...\r"
+                )
+                result.append(wait_line)
         return result
 
     def make_timeout_lines(self, *, get_attempts=ATTEMPTS_BEFORE_TIMEOUT):
         wait_time = self.get_elapsed_time_for_get_attempts(get_attempts)
         adjusted_scenario = Scenario(start_time=wait_time, wait_time_delta=self.wait_time_delta)
         time_out_time = adjusted_scenario.get_time_after_wait()
+        return [f"Exiting after check processing timeout | Check status using: TODO"]
         return [f"{time_out_time} Timed out after {get_attempts} tries."]
 
     def make_outcome_lines(self, get_attempts, *, outcome):
         end_time = self.get_elapsed_time_for_get_attempts(get_attempts)
-        return [f"{end_time} Final status: {outcome}"]
+        return [f"{end_time} Final status: {outcome.title()}"]
 
     def get_elapsed_time_for_get_attempts(self, get_attempts):
         initial_check_time_delta = self.wait_time_delta
-        wait_time_delta = (PROGRESS_CHECK_INTERVAL + self.wait_time_delta) * get_attempts
+        # Extra PROGRESS_CHECK_INTERVAL for the 1-second sleep loop in utils.check_repeatedly,
+        # via make_wait_lines; and extra 2 for the extra (first/last) lines in make_wait_lines.
+        wait_time_delta = (PROGRESS_CHECK_INTERVAL + self.wait_time_delta) * get_attempts + PROGRESS_CHECK_INTERVAL + 2
         elapsed_time_delta = initial_check_time_delta + wait_time_delta
         adjusted_scenario = Scenario(start_time=self.start_time, wait_time_delta=elapsed_time_delta)
         return adjusted_scenario.get_time_after_wait()
@@ -1256,7 +1290,7 @@ class Scenario:
         wait_attempts = get_attempts - 1
         result += scenario.make_uploaded_lines()
         if wait_attempts > 0:
-            result += scenario.make_wait_lines(wait_attempts)
+            result += scenario.make_wait_lines(wait_attempts, outcome=outcome)
         result += scenario.make_outcome_lines(get_attempts, outcome=outcome)
         return result
 
@@ -1273,7 +1307,7 @@ class Scenario:
         scenario = Scenario()
         result = []
         result += scenario.make_uploaded_lines()
-        result += scenario.make_wait_lines(ATTEMPTS_BEFORE_TIMEOUT)
+        result += scenario.make_wait_lines(ATTEMPTS_BEFORE_TIMEOUT - 1, outcome="timeout")
         result += scenario.make_timeout_lines()
         return result
 
@@ -1302,7 +1336,7 @@ def test_submit_any_ingestion_old_protocol():
 
                     assert shown.lines == ["Aborting submission."]
 
-    def mocked_post(url, auth, data, headers, files):
+    def mocked_post(url, auth, data, headers, files, **kwargs):
         # We only expect requests.post to be called on one particular URL, so this definition is very specialized
         # mostly just to check that we're being called on what we think so we can return something highly specific
         # with some degree of confidence. -kmp 6-Sep-2020
@@ -1310,7 +1344,7 @@ def test_submit_any_ingestion_old_protocol():
         assert auth == SOME_AUTH
         ignored(data)
         assert isinstance(files, dict) and 'datafile' in files and isinstance(files['datafile'], io.BytesIO)
-        assert headers == {'Content-type': 'application/json'}
+        assert not headers or headers == {'Content-type': 'application/json'}
         return FakeResponse(201, json={'submission_id': SOME_UUID})
 
     partial_res = {
@@ -1360,7 +1394,7 @@ def test_submit_any_ingestion_old_protocol():
             responses = (partial_res,) * (done_after_n_tries - 1) + (error_res,)
         response_maker = make_alternator(*responses)
 
-        def mocked_get(url, auth):
+        def mocked_get(url, auth, **kwargs):
             print("in mocked_get, url=", url, "auth=", auth)
             assert auth == SOME_AUTH
             if url.endswith("/me?format=json"):
@@ -1456,7 +1490,7 @@ def test_submit_any_ingestion_old_protocol():
                                                             keydict=SOME_KEYDICT,
                                                             upload_folder=None,
                                                             no_query=False,
-                                                            subfolders=False,
+                                                            subfolders=False
                                                         )
         assert shown.lines == Scenario.make_successful_submission_lines(get_request_attempts)
 
@@ -1514,7 +1548,7 @@ def test_submit_any_ingestion_old_protocol():
                                                             keydict=SOME_KEYDICT,
                                                             upload_folder=None,
                                                             no_query=False,
-                                                            subfolders=False,
+                                                            subfolders=False
                                                         )
         assert shown.lines == Scenario.make_successful_submission_lines(get_request_attempts)
 
@@ -1562,7 +1596,7 @@ def test_submit_any_ingestion_old_protocol():
                                                         keydict=SOME_KEYDICT,
                                                         upload_folder=None,
                                                         no_query=True,
-                                                        subfolders=False,
+                                                        subfolders=False
                                                     )
         assert shown.lines == Scenario.make_successful_submission_lines(get_request_attempts)
 
@@ -1616,7 +1650,7 @@ def test_submit_any_ingestion_old_protocol():
 
                                                         assert mock_do_any_uploads.call_count == 0
         assert shown.lines == [
-            "The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.",
+            "The server http://localhost:7777 recognizes you as: J Doe <jdoe@cgap.hms.harvard.edu>",
             "Unsupported Media Type: Request content type multipart/form-data is not 'application/json'",
             "NOTE: This error is known to occur if the server does not support metadata bundle submission."
         ]
@@ -1670,7 +1704,7 @@ def test_submit_any_ingestion_old_protocol():
 
                                                         assert mock_do_any_uploads.call_count == 0
         assert shown.lines == [
-            "The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.",
+            "The server http://localhost:7777 recognizes you as: J Doe <jdoe@cgap.hms.harvard.edu>",
             "Mysterious Error: If I told you, there'd be no mystery.",
         ]
 
@@ -1816,7 +1850,7 @@ def test_submit_any_ingestion_new_protocol():
                                              env=None,
                                              validate_only=False,
                                              no_query=False,
-                                             subfolders=False,)
+                                             subfolders=False)
                     except SystemExit as e:
                         assert e.code == 1
                     else:
@@ -1826,7 +1860,7 @@ def test_submit_any_ingestion_new_protocol():
 
     expect_datafile_for_mocked_post = True
 
-    def mocked_post(url, auth, data=None, json=None, files=None, headers=None):
+    def mocked_post(url, auth, data=None, json=None, files=None, headers=None, **kwargs):
         ignored(data, json)
         content_type = headers and headers.get('Content-type')
         if content_type:
@@ -1920,7 +1954,7 @@ def test_submit_any_ingestion_new_protocol():
             responses = (partial_res,) * (done_after_n_tries - 1) + (error_res,)
         response_maker = make_alternator(*responses)
 
-        def mocked_get(url, auth):
+        def mocked_get(url, auth, **kwargs):
             print("in mocked_get, url=", url, "auth=", auth)
             assert auth == SOME_AUTH
             if url.endswith("/me?format=json"):
@@ -1962,7 +1996,7 @@ def test_submit_any_ingestion_new_protocol():
                                                                      env=None,
                                                                      validate_only=False,
                                                                      no_query=False,
-                                                                     subfolders=False,)
+                                                                     subfolders=False)
                                             except ValueError as e:
                                                 # submit_any_ingestion will raise ValueError if its
                                                 # bundle_filename argument is not the name of a
@@ -2004,8 +2038,7 @@ def test_submit_any_ingestion_new_protocol():
                                                                                  validate_only=False,
                                                                                  upload_folder=None,
                                                                                  no_query=False,
-                                                                                 subfolders=False,
-                                                                                 )
+                                                                                 subfolders=False)
                                                         except SystemExit as e:  # pragma: no cover
                                                             # This is just in case. In fact it's more likely
                                                             # that a normal 'return' not 'exit' was done.
@@ -2018,8 +2051,7 @@ def test_submit_any_ingestion_new_protocol():
                                                             keydict=SOME_KEYDICT,
                                                             upload_folder=None,
                                                             no_query=False,
-                                                            subfolders=False,
-                                                        )
+                                                            subfolders=False)
         assert shown.lines == Scenario.make_successful_submission_lines(get_request_attempts)
 
     dt.reset_datetime()
@@ -2092,8 +2124,7 @@ def test_submit_any_ingestion_new_protocol():
                                                                 keydict=SOME_KEYDICT,
                                                                 upload_folder=None,
                                                                 no_query=False,
-                                                                subfolders=False,
-                                                            )
+                                                                subfolders=False)
         assert shown.lines == Scenario.make_successful_submission_lines(get_request_attempts)
 
     dt.reset_datetime()
@@ -2194,7 +2225,7 @@ def test_submit_any_ingestion_new_protocol():
 
                                                         assert mock_do_any_uploads.call_count == 0
         assert shown.lines == [
-            "The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.",
+            "The server http://localhost:7777 recognizes you as: J Doe <jdoe@cgap.hms.harvard.edu>",
             "Unsupported Media Type: Request content type multipart/form-data is not 'application/json'",
             "NOTE: This error is known to occur if the server does not support metadata bundle submission."
         ]
@@ -2249,7 +2280,7 @@ def test_submit_any_ingestion_new_protocol():
 
                                                         assert mock_do_any_uploads.call_count == 0
         assert shown.lines == [
-            "The server http://localhost:7777 recognizes you as J Doe <jdoe@cgap.hms.harvard.edu>.",
+            "The server http://localhost:7777 recognizes you as: J Doe <jdoe@cgap.hms.harvard.edu>",
             "Mysterious Error: If I told you, there'd be no mystery.",
         ]
 
@@ -2777,7 +2808,7 @@ expected_schema_name = GENERIC_SCHEMA_TYPE
 
 def test_upload_file_to_new_uuid():
 
-    def mocked_execute_prearranged_upload(filename, upload_credentials, auth):
+    def mocked_execute_prearranged_upload(filename, upload_credentials, auth, **kwargs):
         assert filename == mocked_good_filename
         assert upload_credentials == mocked_good_upload_credentials
         assert auth == mocked_good_auth