+++ submit_cgap/tests/test_utils.py
@@ -17,7 +17,8 @@ def shown_output():
         def __init__(self):
             self.lines = []
 
-        def mock_print_handler(self, *args):
+        def mock_print_handler(self, *args, **kwargs):
+            assert kwargs.keys() <= {"end", "flush"}
             text = " ".join(map(str, args))
             print("Mocked print:", text)
             self.lines.append(text)